Add shard test (tps)
